Improved basic phone view for match control

+++ server/tms-utils/src/schemas/scoresheets.rs
@@ -6,7 +6,7 @@ use super::answer::Answer;
 pub struct GameScoresheet {
   team_id: String,
   tournament_id: String,
-  round: u8,
+  round: u32,
   answers: Vec<Answer>,
   private_comment: String,
   public_comment: String
